+++ Assets/AddressableAssetsData/AssetGroups/Schemas/Stage_BundledAssetGroupSchema.asset
@@ -40,6 +40,6 @@ MonoBehaviour:
   m_BundleMode: 0
   m_AssetBundleProviderType:
     m_AssemblyName: Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
-    m_ClassName: CustomAssetBundleProvider
+    m_ClassName: MyCustom.AesCbcAssetBundleProvider
   m_BundleNaming: 0
   m_AssetLoadMode: 0
