+++ Assets/AddressableAssetsData/CustomScripts/AesCbcAssetBundleProvider.cs
@@ -1,0 +1,8 @@
+namespace MyCustom
+{
+    [System.ComponentModel.DisplayName("AES-CBC AssetBundle Provider")]
+    public class AesCbcAssetBundleProvider : CryptoAssetBundleProviderBase
+    {
+        public override ICryptoStreamGetter CryptoStreamGetter => new AesCbcStreamGetter();
+    }
+}
