+++ Assets/AddressableAssetsData/CustomScripts/AesCbcAssetBundleProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be3b6d9fdbb3ee644b7a7973da44bb63
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
