+++ Assets/AddressableAssetsData/CustomScripts/CryptoAssetBundleProviderBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3a39a8f96c1ad749a1dc3cd6dd20489
+guid: 00b27fe07abee98478eb1e69910f7eb3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
