+++ Assets/AddressableAssetsData/CustomScripts/CryptoAssetBundleResource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db143f7a67d927942972aeaf9384881d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
