+++ Assets/AddressableAssetsData/CustomScripts/Editor/BuildScriptEncrypt.asset.meta
