+++ Assets/AddressableAssetsData/CustomScripts/ICryptoStreamGetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb5729d44336b2c4cbef73c4421adf9f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
